requirements.txt
flask
flask-cors
pandas
scikit-learn
numpy
anthropic
python-dotenv
